Adding layouts

+++ src/routes/manage/user/+page.server.ts
@@ -11,7 +11,6 @@ export const load = (async ({cookies}) => {
     const role = cookies.get('session')
     const username = cookies.get('user')
     if (role=='admin'){
-        console.log(username)
         return {
             users
         };
@@ -23,7 +22,7 @@ export const load = (async ({cookies}) => {
 export const actions = {
     add: async ({request}) =>{
         const data = await request.formData()
-        const userdata = {
+        const userData = {
             id:data.get("id") as any,
             username: data.get("username") as string,
             password: data.get("password") as string,
@@ -31,11 +30,12 @@ export const actions = {
             name: data.get("name") as string,
             role: data.get("role") as any
         };
-        const user = new User(userdata)
+        const user = new User(userData)
         console.log(user);
         if(user){
             await db.create(user)
             console.log("User created");
+            console.log(await db.getOne('user', await db.getIdByKey('user','email',user.email)))
             throw redirect(303, '/manage/user')
         }else{
             console.log("enter missing data")
@@ -43,7 +43,7 @@ export const actions = {
     },
     edit: async ({request})=>{
         const data = await request.formData()
-        const userdata = {
+        const userData = {
             id:data.get("id") as any,
             username: data.get("username") as string,
             password: data.get("password") as string,
@@ -51,15 +51,22 @@ export const actions = {
             name: data.get("name") as string,
             role: data.get("role") as any
         };
-        const user = new User(userdata);
-        await db.update(user,userdata.id)
+        const user = new User(userData);
+        await db.update(user,userData.id)
         throw redirect(303, '/manage/user')
     },
     delete: async({request})=>{
         const data = await request.formData()
         const id = data.get("id") as any;
         const user = await db.getOne('user',id);
-        await db.delete(new User(user));
-        throw redirect(303, '/manage/user')
+        if(user.role == 'inactive'){
+            await db.delete(new User(user));
+            throw redirect(303, '/manage/user')
+        }
+        else{
+            throw redirect(303, '/manage/user')
+        }
+
     }
 }satisfies Actions
+
